🚚 extract ingest transaction
