+++ packages/atom.io/internal/src/transaction/index.ts
@@ -4,6 +4,7 @@ export * from "./abort-transaction"
 export * from "./apply-transaction"
 export * from "./build-transaction"
 export * from "./create-transaction"
+export * from "./ingest-transaction-update"
 
 export const TRANSACTION_PHASES = [`idle`, `building`, `applying`] as const
 export type TransactionPhase = (typeof TRANSACTION_PHASES)[number]
